% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ggheader.R
\name{insertGGHeader}
\alias{insertGGHeader}
\title{Inserts ggplot2 setup code.}
\usage{
insertGGHeader()
}
\value{
The ggplot2 code header will be inserted at the cursor position.
}
\description{
\code{insertGGHeader} inserts boilerplate code used in ggplot2 plotting.
}
\details{
This is an RStudio addin, is not interactive, and takes no parameters.
}
\examples{
insertGGHeader()
}

