% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_header.R
\name{insertOccamsHeader}
\alias{insertOccamsHeader}
\title{Inserts boilerplate Occams code header.}
\usage{
insertOccamsHeader()
}
\value{
The Occams code header will be inserted at the cursor position.
}
\description{
\code{insertOccamsHeader} inserts boilerplate Occams code header.
}
\details{
This is an RStudio addin, is not interactive, and takes no parameters.
}
\examples{
insertOccamsHeader()
}

